LayoutTests/imported/w3c/web-platform-tests/streams/readable-byte-streams/detached-buffers.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL detached-buffers.js shared worker wrapper file ReferenceError: Can't find variable: SharedWorker
